Consistent naming and navigation typing.
